Code cleanup (#879)
* change getDelayedState to getBufferedState

* Remove uint64 from blockTime

* change handleRevealState to checkForLastCommitted

* Updated utils/cmd mocks

* CalculateBlockNumberAtEpochBeginning renamed to EstimateBlockNumberAtEpochBeginning

* ClaimStakeReward renamed to ClaimStakerReward

* Update CheckForLastCommitted

* updated GetStatesAllowed to GetFormattedStateNames; removed the comma logic

* getSalt cleanup

* changed int to uint in constants

* Update requested changes

* Update failing test

+++ cmd/initiateWithdraw.go
@@ -102,7 +102,7 @@ func (*UtilsStruct) HandleUnstakeLock(client *ethclient.Client, account types.Ac
 
 	waitFor := big.NewInt(0).Sub(unstakeLock.UnlockAfter, big.NewInt(int64(epoch)))
 	if waitFor.Cmp(big.NewInt(0)) > 0 {
-		timeRemaining := int64(uint32(waitFor.Int64())) * core.EpochLength
+		timeRemaining := uint64(waitFor.Int64()) * core.EpochLength
 		if waitFor.Cmp(big.NewInt(1)) == 0 {
 			log.Infof("Withdrawal period not reached. Cannot withdraw now, please wait for %d epoch! (approximately %s)", waitFor, razorUtils.SecondsToReadableTime(int(timeRemaining)))
 		} else {
